+++ src/avatar/avatar.stories.tsx
@@ -187,6 +187,10 @@ const meta = {
     component: Avatar,
     parameters: {
         badges: ['accessible'],
+        figma: {
+            path: 'Global › Avatar',
+            url: 'https://www.figma.com/design/xo9yAsH8PQUpi0eTJh9pmR/Product-Library---Global?node-id=6957-36532',
+        },
     },
 } satisfies Meta<typeof Avatar>