+++ src/banner/banner.stories.jsx
@@ -388,6 +388,10 @@ export default {
 
     parameters: {
         badges: ['accessible'],
+        figma: {
+            path: 'Web › Components / Todoist › Banner › Banner',
+            url: 'https://www.figma.com/design/LYlWNzvhMDh907l07mPPQk/Product-Library---Web?node-id=15487-102766',
+        },
     },
 }