+++ src/modal/modal.stories.tsx
@@ -32,11 +32,15 @@ import {
 import type { JSX } from 'react'
 
 export default {
-    title: '🪟 Overlays/Modal/Examples',
+    title: '🪟 Overlays/Modal',
     component: ModalComponents.Modal,
     parameters: {
         viewMode: 'story',
         badges: ['accessible'],
+        figma: {
+            path: 'Web › Components / Todoist › Settings › Modal / Header',
+            url: 'https://www.figma.com/design/LYlWNzvhMDh907l07mPPQk/Product-Library---Web?node-id=7443-241153',
+        },
         layout: 'fullscreen',
     },
     decorators: [